aro_agent/aro_agent_backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aro-agent"
version = "0.1.0"
description = "Academic Research Online Agent (arXiv + Crossref + DOAJ)"
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "google-auth",
  "google-auth-oauthlib",
  "google-api-python-client",
  "google-auth-httplib2"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aro_agent*"]
exclude = ["tests*", "out*", "dist*", "build*"]
